Refactoring: Created a module for the Video Panel
Created a module to encapsulate the entire video panel and moved the related components to it.
