+++ src/app/app-routing.module.ts
@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { ChooseVideoComponent } from './choose-video/choose-video.component';
 
 const routes: Routes = [
 ];
