+++ src/app/video-dashboard/choose-video/choose-video.component.scss
