+++ src/app/video-dashboard/language-select/language-select.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { Language } from '../Models/language';
+import { Language } from '../../Models/language';
 import {LANGUAGES} from './Languages';
 
 @Component({
